⬆️ Upgrade @sanity packages to v3
