+++ lib/html_to_portable_text/rules.ts
@@ -1,18 +1,59 @@
-import { deserializeImage } from "./deserializers/image"
-import { deserializeTable } from "./deserializers/table"
+import type { DeserializerRule } from "@sanity/block-tools"
 
-// TODO: remove empty tags
-export const rules = [
+import { getRandomBlockKey } from "./get_random_block_key"
+
+export function isElement(node: Node): node is Element {
+  return node.nodeType === 1
+}
+
+const deserializeTableRows = (el: HTMLTableSectionElement, cellType: string) => {
+  return Array.from(el.children).map((row) => {
+    const cells = Array.from(row.children).map((cell) => ({
+      _type: cellType,
+      text: cell.textContent,
+      _key: getRandomBlockKey(),
+    }))
+
+    return {
+      _type: "tr",
+      cells,
+      _key: getRandomBlockKey(),
+    }
+  })
+}
+
+export const rules: DeserializerRule[] = [
   {
-    deserialize(el: HTMLElement, next: any, block: any) {
-      switch (el.tagName) {
-        case "TABLE":
-          return deserializeTable(el, next, block)
-        case "IMG":
-          return deserializeImage(el, next, block)
-        default:
-          return next()
+    deserialize(el, _next, block) {
+      if (!isElement(el) || el.tagName.toLowerCase() !== "img") {
+        return undefined
       }
+      return block({
+        _type: "image",
+        src: el.getAttribute("src"),
+        alt: el.getAttribute("alt"),
+        height: el.getAttribute("height"),
+        width: el.getAttribute("width"),
+      })
+    },
+  },
+  {
+    deserialize(el, _next, block) {
+      if (!isElement(el) || el.tagName.toLowerCase() !== "table") {
+        return undefined
+      }
+      const thead = el.getElementsByTagName("thead")[0]
+      const tbody = el.getElementsByTagName("tbody")[0]
+
+      const headerRow = thead && deserializeTableRows(thead as HTMLTableSectionElement, "th")
+      const rows = tbody && deserializeTableRows(tbody as HTMLTableSectionElement, "td")
+
+      return block({
+        _type: "table",
+        _key: getRandomBlockKey(),
+        headerRow,
+        rows,
+      })
     },
   },
 ]
