+++ lib/html_to_portable_text/schemas/collibraRichText.ts
@@ -1,13 +1,14 @@
-import Schema from "@sanity/schema"
+import { Schema } from "@sanity/schema"
 
 export const collibraRichText = Schema.compile({
   name: "collibraRichText",
   types: [
     {
-      name: "content",
       type: "object",
+      name: "content",
       fields: [
         {
+          title: "Body",
           name: "body",
           type: "array",
           of: [{ type: "block" }],
