+++ package.json
@@ -50,8 +50,8 @@
   "dependencies": {
     "@azure/app-configuration": "^1.3.1",
     "@portabletext/toolkit": "^2.0.1",
-    "@sanity/block-tools": "^2.35.0",
-    "@sanity/schema": "^2.35.0",
+    "@sanity/block-tools": "^3.8.3",
+    "@sanity/schema": "^3.8.3",
     "axios": "^1.3.5",
     "fp-ts": "^2.13.1",
     "jsdom": "^21.1.1",
